mazebase-training/requirements.txt
backports.shutil-get-terminal-size==1.0.0
-e git+https://github.com/openai/baselines.git@c57528573ea695b19cd03e98dae48f0082fb2b5e#egg=baselines
Bottleneck==1.2.1
cffi==1.10.0
chardet==3.0.3
Click==7.0
cloudpickle==1.2.1
clyent==1.2.2
colorama==0.4.1
ConceptNet==5.7.0
conda==4.3.21
ConfigArgParse==1.0
contextlib2==0.5.5
cryptography==1.8.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
datashape==0.5.4
datrie==0.8
decorator==4.0.11
distributed==1.16.3
docutils==0.13.1
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
fire==0.2.0
Flask==0.12.2
Flask-Cors==3.0.2
ftfy==5.7
future==0.17.1
gym==0.14.0
h5py==2.7.0
HeapDict==1.0.0
html5lib==0.9999999
idna==2.5
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
kaptan==0.5.12
kiwisolver==1.1.0
langcodes==1.4.1
lazy-object-proxy==1.2.2
libtmux==0.8.2
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
marisa-trie==0.7.5
Markdown==3.1.1
MarkupSafe==0.23
matplotlib==2.0.2
# Editable install with no version control (mazebaseenv==0.0.1)
-e /home/kdmarino/Mazebase_gym_environ/mazebaseenv
mistune==0.7.4
mpi4py==3.0.2
mpmath==0.19
msgpack==1.0.0
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.17.0
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
opencv-python==4.1.0.25
openpyxl==2.4.7
ordered-set==3.1.1
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
plotly==3.10.0
ply==3.10
prompt-toolkit==1.0.14
protobuf==3.9.0
psutil==5.2.2
psycopg2-binary==2.8.4
ptyprocess==0.5.1
py==1.4.33
pybullet==2.5.3
pylint==1.6.4
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pyspellchecker==0.5.0
PyYAML==5.1.1
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.12.0
skipthoughts==0.0.1
stable-baselines==2.6.0
tensorboard==1.9.0
termcolor==1.1.0
terminado==0.6
testpath==0.3
tmuxp==1.5.3
toolz==0.8.2
toposort==1.5
torch==1.1.0
torch-vision==0.1.6.dev0
torchfile==0.1.0
torchsummary==1.5.1
torchtext==0.4.0
tornado==4.5.1
tqdm==4.32.2